+++ CHANGELOG.md
@@ -18,7 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     Useful for external QR codes, logos, etc. (#492 #483)
   - update_pcb_characteristics: updates the text you get from *Place* ->
     *Add Board Characteristics*, so you don't need to remove it and place
-    again. (See #384)
+    again. (See #384 #368)
+  - update_stackup: updates the text you get from *Place* ->
+    *Add Stackup Table*, so you don't need to remove it and place
+    again. (See #384 #368)
 - Internal templates:
   - ExportProject: creates a ZIP file containing a self-contained version of
     the project. All footprint, symbols and 3D models are included.